Revert "Changes to share applab page"
