+++ apps/src/StudioApp.js
@@ -26,7 +26,7 @@ var logToCloud = require('./logToCloud');
 var AuthoredHints = require('./authoredHints');
 var Instructions = require('./templates/instructions/Instructions');
 var DialogButtons = require('./templates/DialogButtons');
-var WireframeButtons = require('./templates/WireframeButtons');
+var WireframeSendToPhone = require('./templates/WireframeSendToPhone');
 import InstructionsDialogWrapper from './templates/instructions/InstructionsDialogWrapper';
 import DialogInstructions from './templates/instructions/DialogInstructions';
 var assetsApi = require('./clientApi').assets;
@@ -2029,24 +2029,10 @@ StudioApp.prototype.handleHideSource_ = function (options) {
       } else {
         $(vizColumn).addClass('wireframeShare');
 
-        // Create an empty div on the left for padding
         var div = document.createElement('div');
-        div.className = 'WireframeButtons_container';
-        document.body.insertBefore(div, document.body.firstChild);
-
-        // Add 'withWireframeButtons' class to top level div that wraps app.
-        // This will add necessary styles.
-        div = document.getElementsByClassName('wrapper')[0];
-        if (div) {
-          div.className = 'wrapper withWireframeButtons';
-        }
-
-        // Create div for buttons on the right
-        div = document.createElement('div');
-        div.className = 'WireframeButtons_container';
         document.body.appendChild(div);
         if (!options.level.iframeEmbed) {
-          ReactDOM.render(React.createElement(WireframeButtons, {
+          ReactDOM.render(React.createElement(WireframeSendToPhone, {
             channelId: dashboard.project.getCurrentId(),
             appType: dashboard.project.getStandaloneApp()
           }), div);
