+++ apps/src/code-studio/components/AdvancedShareOptions.jsx
@@ -76,7 +76,7 @@ var AdvancedShareOptions = Radium(React.createClass({
   renderEmbedTab() {
     var url = window.location.href.replace('edit','embed');
     // If you change this width and height, make sure to update the
-    // #visualizationColumn.wireframeShare and .WireframeButtons_container css
+    // #visualizationColumn.wireframeShare css
     var iframeHtml = '<iframe width="352" height="612" style="border: 0px;" src="' +
           url + '"></iframe>';
     return (
